code/requirements.txt
# CBR multi-agent GSM8K experiments
datasets>=2.16.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0
httpx>=0.27.0
tqdm>=4.66.0
python-dotenv>=1.0.0
